Rename to can-element
Closes #22
