--- docs/can-custom-elements.CanElement.md
@@ -1,4 +1,0 @@
-@typedef {function} can-custom-elements.CanElement CanElement
-@parent can-custom-elements.types
-
-An interface for derived elements using either [can-custom-elements] or [can-custom-elements.element].
