+++ docs/can-element.CanElement.md
@@ -1,0 +1,4 @@
+@typedef {function} can-element.CanElement CanElement
+@parent can-element.types
+
+An interface for derived elements using either [can-element] or [can-element.element].
