+++ docs/can-element.CanElement.view.md
@@ -1,10 +1,10 @@
-@typedef {{}} can-custom-elements.CanElement.view CanElement.view
-@parent can-custom-elements.CanElement
+@typedef {{}} can-element.CanElement.view CanElement.view
+@parent can-element.CanElement
 
 A static getter that returns the renderer function used to render the element's shadow DOM.
 
 ```js
-var Element = require("can-custom-elements").Element;
+var Element = require("can-element").Element;
 var view = require("./some-template.stache");
 
 class MyApp extends Element {
