+++ test/attributes-test.js
@@ -1,6 +1,6 @@
 var QUnit = require("steal-qunit");
-var CanElement = require("can-custom-elements").Element;
-var defineAttr = require("can-custom-elements/attributes");
+var CanElement = require("can-element").Element;
+var defineAttr = require("can-element/attributes");
 var compute = require("can-compute");
 var define = require("can-define");
 var stache = require("can-stache");
