+++ test/test.html
@@ -1,4 +1,4 @@
-<title>can-custom-elements</title>
+<title>can-element</title>
 <script>
 	steal = {
 		configDependencies: [
